Add files via upload
The structure of the project is as the follow: all templates (.html) must be inside a folder called "templates", in which will be a layouts folder that holds the .hmtl layouts, all .css or .js files must be inside a folder called "static". The /static and /templates must be siblings, where the flask project should be installed.